requirements.txt
torch>=2.0.0
torchvision>=0.15.0
medmnist>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
Pillow>=9.5.0
jupyter>=1.0.0
ipykernel>=6.25.0
PyYAML>=6.0
